chore: refactor send-deployment type
